Update README.md
With transistor scaling and technology evolution, the exploration of high‐performance and energy‐efficient circuits continues unabated . This search for energy and area‐efficient
circuits has been further enhanced due to the ubiquitous use of
portable devices. To meet the design criteria of modern
high‐performance microprocessors, circuit designers often fall
in a dilemma to make trade‐offs among area, delay, and power
consumption. As a result, logic realization with the least
number of devices and low area with optimal delay and power
has become a major challenge.
The performance of arithmetic circuits determines the
overall performance of a digital processor. The addition
operation plays a fundamental role because several complex
operations are dependent on addition . In digital addition,
the full adder (FA) works as the most important elementary
building block. Hence, the improved design of a FA has to be
pursued vigorously in order to bring overall improvement in
the performance of arithmetic units [6]. Despite having several
existing FA designs, the quest for a new FA design continues to
address the ever‐increasing throughput requirements of modern processors. For these reasons, optimal implementation of
the FA remains a key interest among researchers and VLSI
designers to ensure a high‐performance level in modern digital
systems.
This study presents a full swing hybrid FA based on
XOR‐XNOR gates along with inverters and pass transistors.
The design is simulated in a 16 nm FinFET process to
analyze its performance. Moreover, a performance comparison with existing state of the art FA designs has been carried
out to justify the effectiveness of the XOR‐XNOR‐based FA
presented in this study. The proposed design exhibits high
performance compared to the existing counterparts at the
28 nm technology node across a wide supply voltage range
which makes it quite suitable for state‐of‐the‐art digital
systems.